+++ src/xsl-validator.js
@@ -4,7 +4,7 @@
  */
 
 const {xml, yaml} = require('./helpers')
-const path = require('node:path')
+const path = require('path')
 const {logger} = require('./logger')
 
 /**